chore: move seeds script to src folder
