+++ nest-cli.json
@@ -4,6 +4,9 @@
   "sourceRoot": "src",
   "compilerOptions": {
     "deleteOutDir": true,
-    "builder": "swc"
+    "builder": "swc",
+    "assets": [
+      "prisma/seeds/data/**/*.json"
+    ]
   }
 }
