+++ package.json
@@ -29,7 +29,7 @@
   },
   "prisma": {
     "schema": "prisma/schema.prisma",
-    "seed": "ts-node --compiler-options {\"module\":\"commonjs\"} prisma/seeds/seed.ts"
+    "seed": "node dist/prisma/seeds/seed.js"
   },
   "dependencies": {
     "@nestjs/axios": "^3.1.1",
